rag-agents-demo/requirements.txt
boto3>=1.34.0
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-community>=0.0.20
numpy>=1.24.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.0.0
